More refactoring, create a new intermediate base class
